[feature] msmtp
This add [msmtp](https://marlam.de/msmtp/) in the `businesscard_web` 
container,
allowing the app to use /usr/sbin/sendmail and send emails.

Please note that this is a "quick fix" and the better way to do it would 
be to
use the same service account that is used in the `scripts/extract` file 
with 
the `use Email::Sender::Transport::SMTP;` library.